+++ PCbuild/_freeze_module.vcxproj.filters
@@ -481,6 +481,9 @@
     <ClCompile Include="..\Objects\typeobject.c">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="..\Objects\unicode_format.c">
+      <Filter>Source Files</Filter>
+    </ClCompile>
     <ClCompile Include="..\Objects\unicodectype.c">
       <Filter>Source Files</Filter>
     </ClCompile>
