+++ PCbuild/pythoncore.vcxproj.filters
@@ -1271,6 +1271,9 @@
     <ClCompile Include="..\Objects\typeobject.c">
       <Filter>Objects</Filter>
     </ClCompile>
+    <ClCompile Include="..\Objects\unicode_format.c">
+      <Filter>Objects</Filter>
+    </ClCompile>
     <ClCompile Include="..\Objects\unicodectype.c">
       <Filter>Objects</Filter>
     </ClCompile>
